pyproject.toml
[tool.poetry]
name = "iroha-db-api-gateway"
version = "0.1.0"
description = "Hyperleder Iroha V1 CLi Block Parser and API"
authors = ["Farren Jackson <farren@itbsoftware.co.za>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/"
repository = "https://github.com/"
keywords = ["iroha", "database", "sqlalchemy", "api"]
classifiers = [
    "Programming Language :: Python :: 3.8",
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Development Status :: 2 - Pre-Alpha",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.8"
sqlalchemy = "^1.3.19"
iroha = "^0.0.5"
pandas = "^1.1.1"
psycopg2-binary = "^2.8.5"
huey = "^2.3.0"
gevent = "^20.6.2"
rich = "^6.0.0"
click = "^7.1.2"
flask = "^1.1.2"
flask-sqlalchemy = "^2.4.4"
coverage = "^5.2.1"
alembic = "^1.4.2"
flask_bcrypt = "^0.7.1"
flask_cors = "^3.0.8"
jsonpickle = "^1.4.1"
flask_script = "^2.0.6"
flask_migrate = "^2.5.3"
redis = "^3.5.3"
PyJWT = "^1.7.1"
pydantic = "^1.6.1"
schedule = "^0.6.0"
requests = "^2.24.0"


[tool.poetry.dev-dependencies]
black = "^20.8b1"
bandit = "^1.6.2"
pytest = "^6.0.1"
flake8 = "^3.8.3"
isort = "^5.5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
